requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-js-asset==2.1.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
promise==2.3
pycparser==2.21
PyJWT==1.7.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.1.0
